requirements.txt
Django==4.2.9
djangorestframework==3.14.0
python-dotenv==1.0.0
openai>=2.0.0
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
